prettier and advanced file

+++ contracts/interfaces/IERC20.sol
@@ -1,10 +1,20 @@
 pragma solidity ^0.5.16;
 
+
 interface IERC20 {
-    function approve(address, uint) external;
-    function transfer(address, uint) external;
-    function transferFrom(address, address, uint) external;
-    function deposit() external payable;
-    function withdraw(uint) external;
-    function balanceOf(address) external view returns (uint);
+  function approve(address, uint256) external;
+
+  function transfer(address, uint256) external;
+
+  function transferFrom(
+    address,
+    address,
+    uint256
+  ) external;
+
+  function deposit() external payable;
+
+  function withdraw(uint256) external;
+
+  function balanceOf(address) external view returns (uint256);
 }
